+++ src/public/js/Main.js
@@ -15,22 +15,22 @@ var endangeredTree = L.icon({
 
 /* Map Init || Khởi tạo Map */
 var tiles = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-        maxZoom: 20,
+        maxZoom: 20,    
         attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, Points &copy 2012 LINZ'
     }),
     latlng = L.latLng(10.031517, 105.767851); /*10.031517, 105.767851 vị trí tại B1 CTU */
 
 var map = L.map('map', {
     center: latlng,
-    maxZoom: 20,
+    maxZoom: 19,
     minZoom: 7,
-    zoom: 10,
+    zoom: 17,
     layers: [tiles]
 });
 
 var markers = L.markerClusterGroup({
     chunkedLoading: true,
-    maxClusterRadius: 50,
+    maxClusterRadius: 30,
     spiderfyOnMaxZoom: true,
     showCoverageOnHover: false
 });
@@ -44,14 +44,14 @@ if ('geolocation' in navigator) {
         defY = position.coords.longitude
         var marker = L.marker([defX, defY])
         marker.addTo(map)
-        marker.bindPopup(`Vị trí hiện tại của bạn`, {}) // lat: ${defX.toFixed(3)} long: ${defY.toFixed(3)}
+        marker.bindPopup(`Vị trí hiện tại của gateway`, {}) // lat: ${defX.toFixed(3)} long: ${defY.toFixed(3)}
         marker.openPopup()
     })
 } else {
     console.log('geolocation not available')
 }
 
-//LongLat Converter || Hàm chuyển đổi vị trí
+//LongLat Converter || Hàm chuyển đổi Vĩ độ & Kinh độ
 function handleLongLat(a){
     var x, y, z
     x = parseInt(a / 100)
@@ -65,27 +65,38 @@ async function MapDataFetch() {
     const response = await fetch("/api")
     const data = await response.json()
     for (item of data) {
-        nodeName = item.Name //ID
-        stat = item.Status //Status                     || Trạng thái
-        x = item.ADXL345.X //X
-        y = item.ADXL345.Y //Y
-        z = item.ADXL345.Z //Z
-        description = item.Description //Description    || Chú thích
-        hud = item.Hud //Hudmidity                      || Độ ẩm
-        temp = item.Temp //Temparature                  || Nhiệt độ
-        wind = item.Wind //Wind                         || Gió
-        lat = handleLongLat(item.Coordinate.Latitude) //Latitude       || Vĩ độ
-        lon = handleLongLat(item.Coordinate.Longitude) //Longitude     || Kinh độ
+       
+        nodeName = item.id //ID  
+        stat = item.Trang_thai //Status              || Trạng thái
+        
+        x1 = item.ADXL345_1.X //X1
+        y1 = item.ADXL345_1.Y //Y1
+        z1 = item.ADXL345_1.Z //Z1
+        
+        x2 = item.ADXL345_2.X //X2
+        y2 = item.ADXL345_2.Y //Y2
+        z2 = item.ADXL345_2.Z //Z2
+        
+        description = item.Mo_ta //Description       || Chú thích
+        hud = item.Do_am    //Hudmidity              || Độ ẩm
+        bat = item.Pin      //Battery                || Pin
+        wind = item.Gio     //Wind                   || Gió
+        lat = handleLongLat(item.Toa_do[1]) //Latitude       || Vĩ độ
+        lon = handleLongLat(item.Toa_do[0]) //Longitude     || Kinh độ
+        latest = item.UpdateTime             //Latest update || Lần cập nhật mới nhất
 
         // Check status for the map icon popup || Kiểm tra trạng thái để sửa đổi Icon
-        if (stat === "Good") {
-            stat = "An toàn"
+        wsp = parseFloat(wind)
+        if (stat != "Xanh" && wsp > 50) {
+            stat = "Nguy hiểm"
             var marker = L.marker([lat, lon], {
-                icon: treeIcon
+                icon: endangeredTree
             })
             markers.addLayer(marker)
             map.addLayer(markers)
-        } else if (stat === "Moderate") {
+            alert(`Nguy hiểm tại cây số ${nodeName}`)
+            map.setView([lat, lon], 18)
+        }if (wsp >= 25 && wsp <= 50) {
             stat = "Có nguy cơ"
             var marker = L.marker([lat, lon], {
                 icon: warningTree
@@ -93,9 +104,9 @@ async function MapDataFetch() {
             markers.addLayer(marker)
             map.addLayer(markers)
         } else {
-            stat = "Nguy hiểm"
+            stat = "An toàn"
             var marker = L.marker([lat, lon], {
-                icon: endangeredTree
+                icon: treeIcon
             })
             markers.addLayer(marker)
             map.addLayer(markers)
@@ -106,13 +117,15 @@ async function MapDataFetch() {
 
             ID: ${nodeName}<br>
             Chú thích: ${description}<br>
-            Tọa độ: Lat: ${lat}, Long: ${lon}<br>
-            Độ ẩm: ${hud} %<br>
-            Nhiệt độ: ${temp} &deg;C<br>
-            Gió: ${wind} km/h<br>
-            ADXL345: &nbsp; X: ${x} &nbsp; Y: ${y} &nbsp; Z: ${z}<br>
-            Trạng thái: ${stat}
-
+            Tọa độ: Vĩ độ: ${lat}, Kinh độ: ${lon}<br>
+            Độ ẩm: ${hud}%<br>
+            Pin: ${bat}%<br>
+            Gió: ${wind}km/h<br>
+            ADXL345_Than: &nbsp; X: ${x1}&deg; &nbsp; Y: ${y1}&deg; &nbsp; Z: ${z1}&deg;<br>
+            ADXL345_Goc: &nbsp; X: ${x2}&deg; &nbsp; Y: ${y2}&deg; &nbsp; Z: ${z2}&deg;<br>
+            Trạng thái: ${stat} <br>
+            Thời gian cập nhật: ${latest}
+            
             `);
 
     }
@@ -126,7 +139,7 @@ async function SearchDataFetch() {
     const data = await response.json()
     var data_length = [] //For storing MongoDB object length || Trữ độ dài các đối tượng trong tệp JSON từ MongoDB
     var queryNumber, nodeData
-
+    console.log(data[1])
     //Store the object length for comparation || Lưu độ dài của đối tượng trong tệp JSON
     for (i = 0; i < Object.keys(data).length; i++) {
         data_length.push(`${i+1}`)
@@ -141,14 +154,14 @@ async function SearchDataFetch() {
         queryNumber = Data_type
         console.log(`ID ${Data_type}`)
     } else {
-        alert("Not found")
+        alert("Không tìm thấy!")
         queryNumber = '-1'
     }
 
     parseInt(queryNumber) //Convert to Num || chuyển đổi sang số kiểu int
     queryNumber--
     nodeData = data[queryNumber] //[GET] nodeData at query position || Lấy dữ liệu tại điểm truy vấn q
-    map.setView([nodeData.Coordinate.Latitude, nodeData.Coordinate.Longitude], 15) //Set view to the Node position || Đặt chế độ xem tại vị trí tìm kiếm được
+    map.setView([handleLongLat(nodeData.Toa_do[1]), handleLongLat(nodeData.Toa_do[0])], 20) //Set view to the Node position || Đặt chế độ xem tại vị trí tìm kiếm được
 
 }
 MapDataFetch()
@@ -159,4 +172,4 @@ setInterval(() => {
     MapDataFetch()
     console.log("Map updated!")
     markers.clearLayers()
-}, 5000)
+}, 10000)